refactor(api): unify certification companion alert repository into certification/shared
