+++ api/tests/certification/evaluation/integration/domain/usecases/create-companion-alert_test.js
@@ -1,6 +1,6 @@
 import { usecases } from '../../../../../../src/certification/evaluation/domain/usecases/index.js';
-import * as certificationCompanionAlertRepository from '../../../../../../src/certification/evaluation/infrastructure/repositories/certification-companion-alert-repository.js';
 import { CertificationCompanionLiveAlertStatus } from '../../../../../../src/certification/shared/domain/models/CertificationCompanionLiveAlert.js';
+import * as certificationCompanionAlertRepository from '../../../../../../src/certification/shared/infrastructure/repositories/certification-companion-alert-repository.js';
 import { expect } from '../../../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../../../tooling/databases.js';
